thêm tính năng lựa chọn theo category trong admin
