+++ app/src/main/res/drawable/bg_spinner_dropdown.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <!-- Nền trắng + viền xám + bo góc -->
+    <item>
+        <shape android:shape="rectangle">
+            <solid android:color="@android:color/white" />
+            <stroke android:width="1dp" android:color="#BDBDBD"/>
+            <corners android:radius="8dp"/>
+        </shape>
+    </item>
+
+    <!-- Mũi tên chỉa xuống -->
+    <item
+        android:drawable="@android:drawable/arrow_down_float"
+        android:gravity="center_vertical|end"
+        android:right="12dp"
+    />
+</layer-list>
