+++ app/src/main/res/layout/activity_admin_home.xml
@@ -21,13 +21,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent" />
 
-
-
-
-
-
-
-
     <FrameLayout
         android:id="@+id/fragmentContainer"
         android:layout_width="0dp"
