+++ app/src/main/res/layout/admin_add_menu_item.xml
@@ -1,14 +1,16 @@
-<?xml version="1.0" encoding="utf-8"?>
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    android:fitsSystemWindows="true">
+    android:fitsSystemWindows="true"
+    android:background="@android:color/white"
+    >
 
     <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="wrap_content"
+        android:background="@color/white">
 
         <LinearLayout
             android:id="@+id/layoutForm"
@@ -23,6 +25,7 @@
                 android:text="Thêm món mới"
                 android:textSize="20sp"
                 android:textStyle="bold"
+                android:textColor="#000000"
                 android:layout_marginBottom="16dp" />
 
             <EditText
@@ -32,16 +35,18 @@
                 android:hint="Tên món"
                 android:background="@drawable/edittext_log"
                 android:padding="12dp"
+                android:textColor="#000000"
                 android:layout_marginBottom="12dp" />
 
-            <EditText
-                android:id="@+id/edtCategory"
+            <!-- Spinner thay thế EditText loại món -->
+            <Spinner
+                android:id="@+id/spinnerCategory"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:hint="Loại món"
-                android:background="@drawable/edittext_log"
+                android:background="@drawable/bg_spinner_dropdown"
                 android:padding="12dp"
-                android:layout_marginBottom="12dp" />
+                android:layout_marginBottom="12dp"
+                />
 
             <EditText
                 android:id="@+id/edtPrice"
@@ -51,6 +56,7 @@
                 android:inputType="number"
                 android:background="@drawable/edittext_log"
                 android:padding="12dp"
+                android:textColor="#000000"
                 android:layout_marginBottom="12dp" />
 
             <Button
