+++ app/src/main/res/layout/edit_item.xml
@@ -5,11 +5,13 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    android:fitsSystemWindows="true">
+    android:fitsSystemWindows="true"
+    android:background="@color/white">
 
     <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="wrap_content"
+        android:background="@color/white">
 
         <LinearLayout
             android:id="@+id/layoutForm"
@@ -25,6 +27,7 @@
                 android:text="Chỉnh sửa chi tiết món ăn"
                 android:textSize="20sp"
                 android:textStyle="bold"
+                android:textColor="@color/black"
                 android:layout_marginBottom="16dp" />
 
             <EditText
@@ -34,16 +37,17 @@
                 android:hint="Tên món"
                 android:background="@drawable/edittext_log"
                 android:padding="12dp"
+                android:textColor="@color/black"
                 android:layout_marginBottom="12dp" />
 
-            <EditText
-                android:id="@+id/editCategory"
+            <Spinner
+                android:id="@+id/spinnerEditCategory"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:hint="Loại món"
-                android:background="@drawable/edittext_log"
+                android:background="@drawable/bg_spinner_dropdown"
                 android:padding="12dp"
-                android:layout_marginBottom="12dp" />
+                android:layout_marginBottom="12dp"
+                />
 
             <EditText
                 android:id="@+id/editPrice"
@@ -51,6 +55,7 @@
                 android:layout_height="wrap_content"
                 android:hint="Giá món (đ)"
                 android:inputType="number"
+                android:textColor="@color/black"
                 android:background="@drawable/edittext_log"
                 android:padding="12dp"
                 android:layout_marginBottom="12dp" />
