+++ app/src/main/res/layout/spinner_dropdown_item.xml
@@ -1,0 +1,8 @@
+<TextView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:gravity="center_vertical"
+    android:padding="12dp"
+    android:textColor="#000000"
+    android:background="@color/white"
+    android:textSize="16sp" />
