+++ app/src/main/res/layout/spinner_item.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<TextView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:padding="2dp"
+    android:textColor="#000000"
+    android:textSize="16sp"
+    android:gravity="center_vertical"
+    android:singleLine="true"/>
